@@ -1,56 +1,86 @@
 import { z } from 'zod';
 import { DEFAULT_SETTINGS } from './ultraFeedSettingsTypes';
+import { FeedItemSourceType, allFeedItemSourceTypes } from './ultraFeedTypes';
 
-const ascendingNumbersOrNull = z.array(z.number().positive().nullable()).refine(arr => {
+const sourceWeightsShape = allFeedItemSourceTypes.reduce((acc, key) => {
+  acc[key] = z.number().min(0, { message: "Weight must be non-negative" });
+  return acc;
+}, {} as { [K in FeedItemSourceType]: z.ZodNumber });
+
+const sourceWeightsSchema = z.object(sourceWeightsShape);
+
+const strictlyAscendingNumbersDefinition = (arr: number[]) => {
   let lastNumber: number | undefined;
-  let nullSeen = false;
 
   for (const item of arr) {
-    if (item === null) {
-      nullSeen = true;
-      continue;
-    }
-    if (nullSeen) return false;
     if (lastNumber !== undefined && item <= lastNumber) return false;
     lastNumber = item;
   }
   return true;
-}, {
-  message: "Breakpoints must be ascending positive numbers; the last N values may be unset (null)"
-});
+};
 
-// Schema for the settings object shape that we intend to save
-export const ultraFeedSettingsSchema = z.object({
-  sourceWeights: z.record(
-    z.string(),
-    z.number().min(0, { message: "Weight must be non-negative" })
-  ).default(DEFAULT_SETTINGS.sourceWeights),
-  commentDecayFactor: z.number().positive().default(DEFAULT_SETTINGS.commentDecayFactor),
-  commentDecayBiasHours: z.number().default(DEFAULT_SETTINGS.commentDecayBiasHours),
-  ultraFeedSeenPenalty: z.number()
-    .min(0, { message: "Value must be between 0 and 1" })
-    .max(1, { message: "Value must be between 0 and 1" })
-    .default(DEFAULT_SETTINGS.ultraFeedSeenPenalty),
-  quickTakeBoost: z.number()
-    .min(0.5, { message: "Value must be between 0.5 and 3.0" })
-    .max(3.0, { message: "Value must be between 0.5 and 3.0" })
-    .default(DEFAULT_SETTINGS.quickTakeBoost),
-  threadScoreAggregation: z.enum(['sum', 'max', 'logSum', 'avg'])
-    .default(DEFAULT_SETTINGS.threadScoreAggregation),
-  threadScoreFirstN: z.number().int().positive()
-    .default(DEFAULT_SETTINGS.threadScoreFirstN),
-  incognitoMode: z.boolean().default(DEFAULT_SETTINGS.incognitoMode),
-  postTruncationBreakpoints: ascendingNumbersOrNull
-    .default(DEFAULT_SETTINGS.postTruncationBreakpoints),
+const displaySettingsSchema = z.object({
   lineClampNumberOfLines: z.number().int()
     .min(0, { message: "Value must be between 0 and 10" })
     .max(10, { message: "Value must be between 0 and 10" })
-    .default(DEFAULT_SETTINGS.lineClampNumberOfLines),
-  commentTruncationBreakpoints: ascendingNumbersOrNull
-    .default(DEFAULT_SETTINGS.commentTruncationBreakpoints),
-  postTitlesAreModals: z.boolean().default(DEFAULT_SETTINGS.postTitlesAreModals),
+    .default(DEFAULT_SETTINGS.displaySettings.lineClampNumberOfLines),
+  postTruncationBreakpoints: z.array(z.number().int().min(0))
+    .max(3, { message: "At most 3 post breakpoints allowed" })
+    .refine(strictlyAscendingNumbersDefinition, {
+      message: "Post breakpoints must be non-negative, strictly ascending numbers."
+    })
+    .default(DEFAULT_SETTINGS.displaySettings.postTruncationBreakpoints),
+  commentTruncationBreakpoints: z.array(z.number().int().min(0))
+    .max(3, { message: "At most 3 comment breakpoints allowed" })
+    .refine(strictlyAscendingNumbersDefinition, {
+      message: "Comment breakpoints must be non-negative, strictly ascending numbers."
+    })
+    .default(DEFAULT_SETTINGS.displaySettings.commentTruncationBreakpoints),
+  postTitlesAreModals: z.boolean().default(DEFAULT_SETTINGS.displaySettings.postTitlesAreModals),
+});
 
-}).partial();
+const commentScoringSchema = z.object({
+  commentDecayFactor: z.number().positive({ message: "Must be a positive number" }),
+  commentDecayBiasHours: z.number().min(0, { message: "Must be non-negative" }),
+  ultraFeedSeenPenalty: z.number()
+    .min(0, { message: "Value must be between 0 and 1" })
+    .max(1, { message: "Value must be between 0 and 1" }),
+  quickTakeBoost: z.number()
+    .min(0.5, { message: "Value must be between 0.5 and 3.0" })
+    .max(3.0, { message: "Value must be between 0.5 and 3.0" }),
+  commentSubscribedAuthorMultiplier: z.number()
+    .min(1, { message: "Value must be between 1 and 10" })
+    .max(10, { message: "Value must be between 1 and 10" }),
+  threadScoreAggregation: z.enum(['sum', 'max', 'logSum', 'avg']),
+  threadScoreFirstN: z.number().int().positive({ message: "Must be a positive integer" }),
+});
+
+const threadInterestModelSchema = z.object({
+  commentCoeff: z.number().min(0, { message: "Must be non-negative" }),
+  voteCoeff: z.number().min(0, { message: "Must be non-negative" }),
+  viewCoeff: z.number().min(0, { message: "Must be non-negative" }),
+  onReadPostFactor: z.number().min(0, { message: "Must be non-negative" }),
+  logImpactFactor: z.number(),
+  minOverallMultiplier: z.number().min(0, { message: "Must be non-negative" }),
+  maxOverallMultiplier: z.number().min(0, { message: "Must be non-negative" }),
+}).refine(data => data.minOverallMultiplier <= data.maxOverallMultiplier, {
+  message: "Min overall multiplier must be less than or equal to max overall multiplier",
+  path: ["minOverallMultiplier"], 
+});
+
+const resolverSettingsSchema = z.object({
+  incognitoMode: z.boolean(),
+  sourceWeights: sourceWeightsSchema.default(DEFAULT_SETTINGS.resolverSettings.sourceWeights),
+  commentScoring: commentScoringSchema.default(DEFAULT_SETTINGS.resolverSettings.commentScoring),
+  threadInterestModel: threadInterestModelSchema.default(DEFAULT_SETTINGS.resolverSettings.threadInterestModel),
+});
+
+export const ultraFeedSettingsSchema = z.object({
+  displaySettings: displaySettingsSchema.default(DEFAULT_SETTINGS.displaySettings),
+  resolverSettings: resolverSettingsSchema,
+}).partial()
 
 export type ValidatedUltraFeedSettings = z.infer<typeof ultraFeedSettingsSchema>;
+export type ValidatedCommentScoring = z.infer<typeof commentScoringSchema>;
+export type ValidatedThreadInterestModel = z.infer<typeof threadInterestModelSchema>;
 export type UltraFeedSettingsZodErrors = z.ZodFormattedError<ValidatedUltraFeedSettings, string> | null; 
